+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # node-libzim x.x.x (not yet released)
+* Add support for `getCurrentSize()` method in zimlib's zim::writer.
 
 # node-libzim 0.0.2 (2016-06-22)
 * Initial release.